+++ lib/src/classes/workoutAlerts/cadence_threshold_alert.class.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_workoutkit/workoutkit.dart';
+import 'package:flutter_workoutkit/flutter_workoutkit.dart';
 
 /// Represents a cadence threshold alert
 class CadenceThresholdAlert extends WorkoutAlert {