+++ lib/src/classes/workoutParts/workout_goal.class.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_workoutkit/workoutkit.dart';
+import 'package:flutter_workoutkit/flutter_workoutkit.dart';
 
 /// Represents a workout goal
 class WorkoutGoal {