+++ lib/src/classes/workouts/swim_bike_run_workout.class.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_workoutkit/workoutkit.dart';
+import 'package:flutter_workoutkit/flutter_workoutkit.dart';
 
 /// Represents a swim bike run workout
 class SwimBikeRunWorkout extends Workout {