+++ test/workoutkit_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_workoutkit/workoutkit.dart';
+import 'package:flutter_workoutkit/flutter_workoutkit.dart';
 import 'package:flutter_workoutkit/src/workoutkit_platform_interface.dart';
 import 'package:flutter_workoutkit/src/workoutkit_method_channel.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';